Edit classes, methods for creating, getting, updating, deleting tasks (part6)
